Removed: `IcmpHandler`
